requirements.txt
apify-client
playwright
thefuzz
pandas
python-dotenv
